рефакторит код
